2.x: move DisposableObserver to public area, add some javadocs
