+++ src/main/java/io/reactivex/internal/operators/observable/ObservableDebounce.java
@@ -19,8 +19,7 @@ import io.reactivex.*;
 import io.reactivex.disposables.Disposable;
 import io.reactivex.functions.Function;
 import io.reactivex.internal.disposables.DisposableHelper;
-import io.reactivex.internal.subscribers.observable.DisposableObserver;
-import io.reactivex.observers.SerializedObserver;
+import io.reactivex.observers.*;
 import io.reactivex.plugins.RxJavaPlugins;
 
 public final class ObservableDebounce<T, U> extends AbstractObservableWithUpstream<T, T> {
